requirements.txt
flask
pdfplumber
python-docx
sentence-transformers
faiss-cpu
pandas
numpy
werkzeug
